Commit initial Chat-Proxy.js as index.js
A bunch of changes relating to the initial commit of index.js.

The package.json file changed to get npm install to work. A few
more changes to try to fix unit tests but gave up.

The aws-sdk is made a dev-only dependency in package.json.

Before this, the deployment package was too large for me to view the
code. On examination I found this:
https://docs.aws.amazon.com/lambda/latest/dg/nodejs-create-deployment-pkg.html

"Note The Lambda service has preinstalled the AWS SDK for Node.js."

Note also that the Handler had to be changed to match the function
name. I chose index.js as it seems like SAM or Lambda renames it as
index.js if it's anything else.

The change is based on the Stack Overflow answer here:
https://stackoverflow.com/a/30672061/3787051